+++ test/select_form_field_test.dart
@@ -4,10 +4,16 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:select_form_field/select_form_field.dart';
 
 void main() {
-  testWidgets('Instantiate SelectFormField', (WidgetTester tester) async {
+  testWidgets('Testing instantiate SelectFormField',
+      (WidgetTester tester) async {
     var myWidget = MyWidget();
     await tester.pumpWidget(myWidget);
 
+    expect(
+      find.text('test'),
+      findsOneWidget,
+      reason: 'SelectFormField value not found!',
+    );
     var selectField = find.byType(SelectFormField);
     expect(
       selectField,
@@ -17,26 +23,46 @@ void main() {
     expect(
       find.text('Circle Label'),
       findsOneWidget,
-      reason: 'Initial label value wrong!',
+      reason: 'SelectFormField initial label value not found!',
     );
+
     await tester.tap(selectField);
-    await tester.pump();
+    await tester.pumpAndSettle();
+
     var boxLabel = find.text('Box Label');
     expect(
       boxLabel,
       findsOneWidget,
-      reason: 'Box option wrong!',
+      reason: 'SelectFormField Box option not displayed on dialog!',
     );
     expect(
       find.text('Star Label'),
       findsOneWidget,
-      reason: 'Star option wrong!',
+      reason: 'SelectFormField Star option not displayed on dialog!',
     );
-    await tester.tap(boxLabel);
+
+    var item = find.ancestor(
+      of: boxLabel,
+      matching: find.byType(InkWell),
+    );
+    expect(
+      item,
+      findsOneWidget,
+      reason: 'SelectFormField InkWell not found!',
+    );
+
+    await tester.tap(item);
+    await tester.pumpAndSettle();
+
     expect(
       find.text('Box Label'),
       findsOneWidget,
-      reason: 'Box not selected!',
+      reason: 'SelectFormField Box not selected!',
+    );
+    expect(
+      find.text('boxValue'),
+      findsOneWidget,
+      reason: 'SelectFormField value result not changed!',
     );
   });
 }
@@ -49,7 +75,7 @@ class MyWidget extends StatefulWidget {
 }
 
 class _MyWidgetState extends State<MyWidget> {
-  String _value = '';
+  String _value = 'test';
 
   final List<Map<String, dynamic>> _items = [
     {
@@ -83,7 +109,6 @@ class _MyWidgetState extends State<MyWidget> {
               labelText: 'Shape',
               items: _items,
               onChanged: (val) => setState(() => _value = val),
-              onSaved: (val) => setState(() => _value = val),
             ),
             Text(_value),
           ],