+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.0.0 - 2020-07-17
+
+* Fixed use of widget with empty initial value
+* Improved README.md
+* Improved testing
+* Improved example app
+* Final validation
+
 ## 0.1.1 - 2020-07-16
 
 * Example page published